Fixed compile warning
<==== Mercurial data ====>
Branch: cloud_18.2_backend

+++ nx_cloud/backend/nx_data_sync_engine/src/nx/data_sync_engine/statistics/provider.cpp
@@ -48,7 +48,7 @@ Statistics Provider::statistics() const
     const auto connections = m_connectionManager.getConnections();
 
     Statistics data;
-    data.connectedServerCount = connections.size();
+    data.connectedServerCount = static_cast<int>(connections.size());
     for (const auto& connection: connections)
         ++data.connectedServerCountByVersion[connection.userAgent];
